Start supporting the MasterPlus+ protocol for the CM RGB Controller device.
Note that this version of the protocol is missing a lot of functionality compared to the other protocol (the RGB LED Controller one). It is, however, much simpler so I'm starting with this and we'll go from there.

Also worth noting that I don't know a thing about USB and have just guessed at how to interpret the captures and turn them into buffers in C++. Once I get some confirmation that the way I'm building these packets will result in what we see in the captures then I'll try uncommenting the USB calls and see if I can control the device with it.