Read mode config from device on startup
For all the simple modes we just query the device, but the "Multiple" mode has
per-LED color and if the device doesn't start up in that mode then there's
nowhere to store those per-LED colors. So there's a special case in SetMode()
which populates the LED colors for per-LED mode.

Also start implementing a virtual "direct" mode. Since this device does not
have a "direct" mode instead I'm doing it in software and either saving or
not saving the changes depending on the mode.