Use longer waiting time when receiving data from wireless Razer devices
For wireless Razer devices we need to give the device more time to respond,
before data is properly sent back to us. Without this extended waiting time
the tested devices (Viper Ultimate, BlackWidow v3 Mini) the device will respond
with the apropriate report, but the arguments will be all 0's (hence why the
serial number of the e.g. the Viper Ultimate would read properly when the device
is wired, but be an empty string when wireless).