Allow controller update flags to be reset before or after the update
While updating a controller the update flags were reset after the update was
finished. This might swallow update calls when a device takes a little bit of
time to update, while the user requests some more update, e.g. by changing the
brightness. On the other hand resetting before the update broke some devices
in other ways. So this commit introduces a way for each controller to choose,
wheter it wants to be reset before or after an update.